Install Karma, Jasmine, etc
Some of the functions in app.js look a little hairy, and it'd be nice to
refactor them but I'm scared of doing that without some test cases.
Trying to make all this test runner stuff seemed a bit magical and made
me sad. I don't understand why setting up simple isolated tests needs to
be this complicated. Oh well ¯\_(ツ)_/¯